update shared folder
