+++ node/shared/errors/Messages.ts
@@ -1,0 +1,5 @@
+export const Messages: {
+    [key: string]: string | ((...args: string[]) => string)
+} = {
+    SEQUENCE_EMPTY: (name: string) => `Sequence '${name}' is empty`
+}
