+++ node/shared/errors/index.ts
@@ -1,0 +1,1 @@
+export * from './Error.js'
